\name{getFittedTable-methods}
\docType{methods}
\alias{getFittedTable-methods}
\alias{getFittedTable,SquadSimResServiceImpl-method}
\title{Build the table of fitted values}
\description{
This method creates a table containing the interpolated/fitted value of the simulation curve at the user-defined time point. Each column represents the condition and each line, the component node. It is used as basis for the Correlation Circle and the Prediction Map.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "SquadSimResServiceImpl")}}{
 See \code{\link{SquadSimResServiceImpl-class}} for more information about object construction.
}
}}
\keyword{methods}
\keyword{fitted}
\keyword{interpolated}