\name{plotCC-methods}
\docType{methods}
\alias{plotCC-methods}
\alias{plotCC,SquadSimResServiceImpl-method}
\title{plot the correlation circle}
\description{
This method generates the correlation circle. Correlation circle is used to assess the coherence of a model. Indeed, each vector represents our conditions and the angle between each vector give the correlation between the variables. if the angle is higher than 90 degree there is no correlation.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "SquadSimResServiceImpl")}}{
 See \code{\link{SquadSimResServiceImpl-class}} for more information about object construction.
}
}}
\keyword{methods}
\keyword{correlation}
\keyword{PCA}