\name{plotPredMap-methods}
\docType{methods}
\alias{plotPredMap-methods}
\alias{plotPredMap,SquadSimResServiceImpl-method}
\title{Display a prediction map}
\description{
This method is used to visualize the prediction heat-map. The prediction map is generated by mapping a color code on the ratio of the fitted values. 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "SquadSimResServiceImpl")}}{
 See \code{\link{SquadSimResServiceImpl-class}} for more information about object construction.
}
}}
\keyword{methods}
\keyword{prediction}
