\name{plotPredMap}
\alias{plotPredMap}
\title{Prediction map}
\description{
This method is used to visualize the prediction heat-map. The prediction map is generated by mapping a color code on the ratio of the fitted values. 
}
\usage{
plotPredMap(object)
}
\arguments{
  \item{object}{
	object of the Class \code{SquadSimResServiceImpl}.
}
}
\details{
This method firstly interpolate the SQUAD Simulation Result. Then, it performs a ratio at a user-defined time point. Finally, a color code is assigned to the ratio value. The darker the color, the higher the increase of the node activation.
}
\author{
Martial Sankar
}

\seealso{
More information about heatmap, \code{\link{heatmap}}
More specific information in \code{\link{plotPredMap-methods}}.
For more informations about the input object see \code{\link{SquadSimResServiceImpl-class}}.
}


\examples{
fpath <- system.file("extdata", package="SQUADD")
fileModel <- file.path(fpath,"data_IAA")
nm <- c("ARF(a)", "ARF(i)", "AR_Genes", "Aux/IAA", "BES1/BZR1",
 "BIN2", "BR", "BRI1-BAK1","BRR_Genes","BRX","BR_Biosynthesis",  "BZR1", "DO",
  "IAA", "IAA_Biosynthesis",   "NGA1", "PIN", "SCFTir1","StimAux", "StimBR")
t <- 50

## call constructor
obj <- simResService(
		folder=fileModel,
		time=t,
		ncolor=5,
		legend=nm,
		indexDeno=1,
		method="lowess")

## Apply method 
plotPredMap(obj)

}

\keyword{prediction}
\keyword{heatmap}
