\name{plotSimMatrix-methods}
\docType{methods}
\alias{plotSimMatrix-methods}
\alias{plotSimMatrix,SquadSimResServiceImpl-method}
\title{Generate a SQUAD simulation matrix}
\description{
This method is used to create a simulation matrix. A simulation matrix permits to follow the activation signal of an user-defined number of Node. Each columns correspond to the user-selected component node and the each rows, to the conditions.
(more details in \code{\link{plotSimMatrix}})}
\section{Methods}{
\describe{

\item{\code{signature(object = "SquadSimResServiceImpl")}}{
 See \code{\link{SquadSimResServiceImpl-class}} for more information about object construction.
}
}}
\keyword{methods}
\keyword{simulation}
