\name{SQUADD-package}
\alias{SQUADD-package}
\alias{SQUADD}
\docType{package}
\title{Collection of visualization tools to perform a SQUAD-based quantitative analysis 
~~SQUADD~~}
\description{
The SQUAD (Standardized Qualitative Dynamical Systems) software is dedicated to the analysis of boolean logical model(Di Cara et al 2007). It was the first tool to combine the traditional boolean approach and a continuous system. This novel feature permits the dynamical analysis of regulatory network in the absence of kinetic data. In order to get the best of this approach, we provide here a R package, SQUADD, a SQUAD add-on for : (i) simplify the visualization of the initial SQUAD simulation results. (ii) generate qualitative predictions of the component change between two conditions. (iii) Assess the coherence of a model by PCA analysis.
}
\details{
\tabular{ll}{
Package: \tab SQUADD\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2010-06-14\cr
License: \tab GPL(>=2)\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 2.11.0)\cr
Imports: \tab graphics, grDevices, methods, RColorBrewer, stats, utils\cr
}
}
\author{
Martial Sankar  <martial.sankar@unil.ch>
}
\references{
Di Cara A, Garg A, De Micheli G, Xenarios I, Mendoza L (2007) Dynamic simulation of regulatory networks using SQUAD. BMC Bioinformatics 8: 462.
Dominik M Wittmann, Jan Krumsiek, Julio Saez-Rodriguez, Douglas A Lauffenburger, Steffen Klamt and Fabian J Theis (2009) Transforming Boolean models to continuous models: methodology and application to T-cell receptor signaling. PLOS Computational Biology 3: 98
}
\keyword{ package }

